fix: separated event handler wrapper
